WIP on catalog page
