+++ app/catalog-tab/index/controller.js
@@ -1,10 +1,13 @@
 import Ember from 'ember';
 import { isAlternate } from 'ui/utils/platform';
+import C from 'ui/utils/constants';
+import { getCatalogSubtree } from 'ui/utils/parse-catalog-setting';
 
 export default Ember.Controller.extend({
   application: Ember.inject.controller(),
   catalog: Ember.inject.service(),
   settings: Ember.inject.service(),
+  projectId: Ember.computed.alias(`tab-session.${C.TABSESSION.PROJECT}`),
 
   catalogController: Ember.inject.controller('catalog-tab'),
   category: Ember.computed.alias('catalogController.category'),
@@ -38,8 +41,9 @@ export default Ember.Controller.extend({
     }
   },
 
-  init() {
-  },
+  filters: Ember.computed(`settings.${C.SETTING.CATALOG_URL}`, function() {
+    return getCatalogSubtree(this.get(`settings.${C.SETTING.CATALOG_URL}`), this.get('projectId'));
+  }),
 
   arrangedContent: Ember.computed('model.catalog', 'search', function() {
     var search = this.get('search').toUpperCase();
