+++ app/catalog-tab/index/template.hbs
@@ -2,67 +2,85 @@
   <h1>{{t 'catalogPage.index.header'}} <span class="text-capitalize">{{catalogId}}</span></h1>
 
   <div class="pull-right">
-    <button type="button" class="btn bg-default btn-sm" disabled={{eq updating 'yes'}} {{action "update"}}>
+    <button type="button" class="btn bg-primary btn-sm" disabled={{eq updating 'yes'}} {{action "update"}}>
       {{#if (eq updating "yes")}}
         <i class="icon icon-spinner icon-spin"></i>
       {{else}}
         {{#if (eq updating "error")}}
           <i class="icon icon-spinner icon-alert"></i>
         {{else}}
+          {{t 'catalogPage.index.refreshBtn'}}
           <i class="icon icon-refresh"></i>
         {{/if}}
       {{/if}}
     </button>
   </div>
 
-  <div class="btn-group pull-right ml-10">
-    <button type="button" class="btn bg-default btn-sm dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">{{t 'catalogPage.index.category'}} <span class="text-capitalize">{{category}}</span> <i class="icon icon-chevron-down"></i></button>
-    <ul class="dropdown-menu dropdown-menu-right">
-      {{#each categories as |opt|}}
-        {{#if (eq opt "all")}}
-          <li class="text-capitalize {{if (eq category opt) 'active'}}">{{#link-to parentRoute (query-params category=opt)}}{{opt}}{{/link-to}}</li>
-          <li class="divider"></li>
-        {{else}}
-          <li class="{{if (eq category opt) 'active'}}">
-            {{#link-to parentRoute (query-params category=opt)}}
-              {{opt}}
-            {{/link-to}}
-          </li>
-        {{/if}}
-      {{/each}}
-    </ul>
-  </div>
 
-  <div class="btn-group searchable pull-right">
-    <i class="icon icon-search"></i>
-    {{input placeholder=(t 'catalogPage.index.search.placeholder') class="form-control input-sm" value=search}}
-  </div>
 </section>
 
 <section class="pl-15 pr-15 clearfix">
-  {{#each arrangedContent as |catalogItem|}}
-    {{#catalog-box model=catalogItem showSource=showCatalogDropdown as |section|}}
-      {{#if (eq section 'footer')}}
-        {{#if (eq catalogItem.templateBase 'infra')}}
-          {{#if catalogItem.exists}}
-            <button type="button" class="btn btn-sm bg-default bg-disabled" {{action "launch" catalogItem.id true allowedKeys="meta ctrl"}}>{{t 'catalogPage.index.action.alreadyDeployed'}}</button>
-          {{else if catalogItem.supported}}
-            <button class="btn btn-sm bg-primary" {{action "launch" catalogItem.id}}>{{t 'catalogPage.index.action.launch'}}</button>
-          {{else}}
-            <button type="button" class="btn btn-sm bg-default bg-disabled" {{action "launch" catalogItem.id true allowedKeys="meta ctrl"}}>{{t 'catalogPage.index.action.notCompatible'}}</button>
+  <div class="row">
+    <div class="col span-3 catalog-search">
+      <div class="mb-15">
+        <label>{{t 'catalogPage.index.search.label'}}</label>
+        <div class="input-group full-width">
+          {{input value=search type="search" placeholder=(t 'generic.search')}}
+          {{#if search}}
+            <span class="input-group-btn">
+              <button class="btn bg-primary" {{action 'clearSearch'}}><i class="icon icon-close"/></button>
+            </span>
           {{/if}}
-        {{else}}
-          <button type="button" class="btn btn-sm bg-primary" {{action "launch" catalogItem.id}}>{{t 'catalogPage.index.action.launch'}}</button>
-        {{/if}}
-      {{/if}}
-    {{/catalog-box}}
-  {{else}}
-    <div class="text-muted mt-20">
-      {{#if showCatalogDropdown}}
-        {{t 'catalogPage.index.noData.plural'}}
-      {{else}}
-        {{t 'catalogPage.index.noData.singular'}}
-      {{/if}}
+        </div>
+      </div>
+      <div>
+        <label>{{t 'catalogPage.index.filters.label'}}</label>
+        <div class="dropdown btn-group filter-group">
+          <button type="button" class="btn bg-default btn-md dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" style="text-align: left;"><i class="icon icon-chevron-down pull-right"></i><span class="text-capitalize">{{catalogId}}</span></button>
+          <ul class="dropdown-menu dropdown-menu-right">
+            {{#each filters as |opt|}}
+              <li class="text-capitalize {{if (eq catalogId opt.queryParams.catalogId) 'active'}}">
+                {{#link-to parentRoute (query-params catalogId=opt.queryParams.catalogId)}}
+                  <i class="{{opt.icon}}"></i>
+                  {{#if opt.localizedLabel}}
+                    {{t opt.localizedLabel}}
+                  {{else}}
+                    {{opt.label}}
+                  {{/if}}
+                {{/link-to}}
+              </li>
+            {{/each}}
+          </ul>
+
+        </div>
+      </div>
     </div>
-  {{/each}}
+    <div class="col span-9">
+      {{#each arrangedContent as |catalogItem|}}
+        {{#catalog-box model=catalogItem showSource=showCatalogDropdown as |section|}}
+          {{#if (eq section 'footer')}}
+            {{#if (eq catalogItem.templateBase 'infra')}}
+              {{#if catalogItem.exists}}
+                <button type="button" class="btn btn-sm bg-default bg-disabled" {{action "launch" catalogItem.id true allowedKeys="meta ctrl"}}>{{t 'catalogPage.index.action.alreadyDeployed'}}</button>
+              {{else if catalogItem.supported}}
+                <button class="btn btn-sm bg-primary" {{action "launch" catalogItem.id}}>{{t 'catalogPage.index.action.launch'}}</button>
+              {{else}}
+                <button type="button" class="btn btn-sm bg-default bg-disabled" {{action "launch" catalogItem.id true allowedKeys="meta ctrl"}}>{{t 'catalogPage.index.action.notCompatible'}}</button>
+              {{/if}}
+            {{else}}
+              <button type="button" class="btn btn-sm bg-primary" {{action "launch" catalogItem.id}}>{{t 'catalogPage.index.action.launch'}}</button>
+            {{/if}}
+          {{/if}}
+        {{/catalog-box}}
+      {{else}}
+        <div class="text-muted mt-20">
+          {{#if showCatalogDropdown}}
+            {{t 'catalogPage.index.noData.plural'}}
+          {{else}}
+            {{t 'catalogPage.index.noData.singular'}}
+          {{/if}}
+        </div>
+      {{/each}}
+    </div>
+  </div>
 </section>
