+++ app/styles/base/_helpers.scss
@@ -114,6 +114,10 @@
   text-transform: uppercase;
 }
 
+.text-capitalize {
+  text-transform: capitalize;
+}
+
 .hide {
   display: none !important;
 }
