+++ app/styles/layout/_loading.scss
@@ -11,6 +11,8 @@
   .lid:after,
   .lid,
   .square,
+  .square:before,
+  .square:after,
   .icon-spin {
     animation: none;
     -webkit-animation:: none;
