+++ app/styles/pages/_catalog.scss
@@ -13,6 +13,30 @@
   }
 }
 
+.catalog-search {
+  margin-top: 28px;
+  padding: 20px;
+  background-color: $accent-one;
+
+  .dropdown {
+    .btn {
+      background-color: $input-bg;
+      i {
+        margin-top: 2px;
+      }
+    }
+  }
+  input[type=text] {
+    // background-color: white;
+  }
+}
+.filter-group {
+  > .dropdown-toggle {
+    // background-color: white;
+    width: 100%;
+  }
+}
+
 .line-numbers {
   &.constrained {
     min-height:200px;
