+++ app/styles/themes/_dark.scss
@@ -91,7 +91,7 @@ $component-active-bg                 : $primary;
 $input-color-placeholder             : darken($light-grey, 50%);
 
 //** `<input>` background color
-$input-bg                            : rgba($light-grey, .1);
+$input-bg                            : $light-grey;
 $input-bg-hover                      : $light-grey;
