+++ app/styles/themes/_light.scss
@@ -87,12 +87,12 @@ $component-active-bg                 : $primary;
 
 
 
-// Forms 
+// Forms
 // Placeholder text color
 $input-color-placeholder             : $mid-grey;
 
 // `<input>` background color
-$input-bg                            : rgba($light-grey, .2);
+$input-bg                            : lighten($light-grey, 4%);
 $input-bg-hover                      : $mid-grey;
 
 // `<input disabled>` background color
