+++ translations/en-us.yaml
@@ -385,8 +385,12 @@ backupTargetsPage:
 catalogPage:
   index:
     header: 'Catalog:'
+    refreshBtn: Refresh
     search:
+      label: Search
       placeholder: Search...
+    filters:
+      label: Catalog Filter
     category: "Category:"
     certified:
       rancher:
